refactor(launcher): extract main.rs logic into sibling modules
