+++ .changesets/1782974746-refactor-launcher-extract-main-rs-logic-into-sibling-modules-d0a3.md
@@ -1,0 +1,5 @@
+---
+type: patch
+---
+
+refactor(launcher): extract main.rs logic into sibling modules
