+++ agentmux-launcher/src/binary_resolution.rs
@@ -1,0 +1,47 @@
+// Copyright 2025-2026, AgentMux Corp.
+// SPDX-License-Identifier: Apache-2.0
+
+/// Find the CEF host binary in the runtime directory.
+/// Tries versioned name first (agentmux-X.Y.Z.exe), then the old
+/// agentmux-cef-X.Y.Z.exe pattern for backwards compat, then plain
+/// agentmux-cef.exe (dev mode).
+pub(crate) fn find_cef_binary(runtime_dir: &std::path::Path) -> std::path::PathBuf {
+    let ext = if cfg!(target_os = "windows") { ".exe" } else { "" };
+
+    let versioned = format!("agentmux-{}{}", env!("CARGO_PKG_VERSION"), ext);
+    let versioned_path = runtime_dir.join(&versioned);
+    if versioned_path.exists() {
+        return versioned_path;
+    }
+
+    if let Ok(entries) = std::fs::read_dir(runtime_dir) {
+        let prefix = "agentmux-";
+        let cef_prefix = "agentmux-cef";
+        for entry in entries.flatten() {
+            let name = entry.file_name();
+            let name = name.to_string_lossy();
+            if name.starts_with(prefix)
+                && !name.starts_with(cef_prefix)
+                && !name.starts_with("agentmux-srv")
+                // CRITICAL for the flat dev layout (macOS/Linux Phase 1):
+                // launcher + host + srv share one dir, so the launcher
+                // binary itself matches `agentmux-*`. Without this guard
+                // the launcher resolves ITSELF as the host and spawns a
+                // recursive launcher fork bomb. On Windows the launcher
+                // lives at the root (not in runtime/), so this is a no-op.
+                && !name.starts_with("agentmux-launcher")
+                && name.ends_with(ext)
+            {
+                return entry.path();
+            }
+        }
+    }
+
+    let versioned_old = format!("agentmux-cef-{}{}", env!("CARGO_PKG_VERSION"), ext);
+    let versioned_old_path = runtime_dir.join(&versioned_old);
+    if versioned_old_path.exists() {
+        return versioned_old_path;
+    }
+
+    runtime_dir.join(format!("agentmux-cef{}", ext))
+}
