+++ agentmux-launcher/src/host_spawn.rs
@@ -1,0 +1,260 @@
+// Copyright 2025-2026, AgentMux Corp.
+// SPDX-License-Identifier: Apache-2.0
+
+use crate::logging::log;
+
+/// Spawn the CEF host suspended, assign it to the launcher's Job Object, and
+/// resume it. Returns the running child, or `None` if any step failed — the
+/// caller decides (fatal on first launch, give-up on a restart). `splash_event`
+/// is passed on every launch — including restarts — so a relaunched host can
+/// still dismiss a splash left pending by a host that crashed pre-first-frame.
+/// `disable_gpu` is the retry ladder's rung-2 degraded mode (spec §7): when set
+/// the host is launched with `--disable-gpu` (software rendering).
+#[cfg(target_os = "windows")]
+pub(crate) fn spawn_host_supervised(
+    real_exe: &std::path::Path,
+    args: &[String],
+    srv: &crate::srv_spawner::SrvSpawnResult,
+    host_env: &[(&'static str, std::ffi::OsString)],
+    pipe_path: &str,
+    job_present: bool,
+    job_handle: windows_sys::Win32::Foundation::HANDLE,
+    splash_event: Option<&str>,
+    disable_gpu: bool,
+) -> Option<tokio::process::Child> {
+    use windows_sys::Win32::System::Threading::CREATE_SUSPENDED;
+
+    // The launcher's resolved `real_exe` lives in `runtime/`; pass its
+    // parent dir as AGENTMUX_HOME so the host can anchor asset lookups
+    // (frontend/index.html, etc.) on something stable rather than
+    // `std::env::current_exe()`. Windows's GetModuleFileName keeps
+    // returning the original load-time path even after the parent dir
+    // is renamed or unlinked out from under it — the 2026-05-28
+    // incident pattern. AGENTMUX_HOME is whatever path *we* (the
+    // launcher) successfully resolved real_exe through, so it always
+    // points at the runtime dir that actually contains the binaries.
+    // See docs/retro/retro-portable-rm-running-install-2026-05-28.md.
+    let host_runtime_dir = real_exe.parent().map(|p| p.to_path_buf());
+
+    let mut host_cmd = tokio::process::Command::new(real_exe);
+    host_cmd
+        .args(args)
+        .env("AGENTMUX_BACKEND_WS", &srv.ws_endpoint)
+        .env("AGENTMUX_BACKEND_WEB", &srv.web_endpoint)
+        .env("AGENTMUX_BACKEND_PID", srv.pid.to_string())
+        .env("AGENTMUX_PENDING_MIGRATIONS", srv.pending_migrations.to_string())
+        .env("AGENTMUX_AUTH_KEY", &srv.auth_key)
+        .env("AGENTMUX_INSTANCE_ID", &srv.instance_id)
+        .envs(host_env.iter().cloned())
+        .env("AGENTMUX_LAUNCHER_PIPE", pipe_path)
+        .creation_flags(CREATE_SUSPENDED)
+        .kill_on_drop(false); // J0 handles cleanup.
+    if let Some(dir) = host_runtime_dir {
+        host_cmd.env("AGENTMUX_HOME", dir);
+    }
+    if let Some(name) = splash_event {
+        host_cmd.env("AGENTMUX_SPLASH_EVENT", name);
+    }
+    // Retry-ladder rung 2 (spec §7): software rendering — no GPU process to
+    // crash. A Chromium switch the host forwards to CEF.
+    if disable_gpu {
+        host_cmd.arg("--disable-gpu");
+    }
+
+    let mut host_child = match host_cmd.spawn() {
+        Ok(c) => c,
+        Err(e) => {
+            log(&format!("failed to spawn CEF host: {}", e));
+            return None;
+        }
+    };
+    let host_pid = host_child.id().unwrap_or(0);
+    log(&format!("spawned CEF host pid={} (suspended)", host_pid));
+
+    // Assign to J0 BEFORE resuming so CEF render children inherit the job.
+    if job_present && host_pid != 0 {
+        match crate::srv_spawner::assign_pid_to_job(host_pid, job_handle) {
+            Ok(()) => log(&format!(
+                "Job Object assigned to host pid={}, KILL_ON_JOB_CLOSE active",
+                host_pid
+            )),
+            Err(e) => log(&format!(
+                "WARN: AssignProcessToJobObject(host) failed: {} — host children may escape job",
+                e
+            )),
+        }
+    }
+
+    // Resume the suspended main thread.
+    if let Err(e) = resume_main_thread(host_pid) {
+        log(&format!("failed to resume host pid={}: {}", host_pid, e));
+        let _ = host_child.start_kill();
+        return None;
+    }
+    Some(host_child)
+}
+
+/// Spawn the CEF host on Unix with the srv endpoints + canonical
+/// data-dir env handed off, mirroring `spawn_host_supervised` minus the
+/// Windows-only machinery (Job Object assignment, CREATE_SUSPENDED /
+/// resume, named-pipe handle, splash event). `disable_gpu` is the retry
+/// ladder's degraded rung (software rendering). Returns the running
+/// child or `None` if spawn failed.
+///
+/// AGENTMUX_HOME is intentionally NOT set: on the flat dev layout the
+/// host's `current_exe().parent()` fallback resolves to the same dir the
+/// launcher would export, so omitting it keeps asset + framework lookup
+/// byte-identical to the legacy direct-invoke (`task dev:standalone`)
+/// path. Phase 2 will set it once the production runtime/ layout lands
+/// on macOS.
+#[cfg(not(target_os = "windows"))]
+pub(crate) fn spawn_host_unix(
+    real_exe: &std::path::Path,
+    args: &[String],
+    srv: &crate::srv_spawner::SrvSpawnResult,
+    host_env: &[(&'static str, std::ffi::OsString)],
+    disable_gpu: bool,
+) -> Option<tokio::process::Child> {
+    let mut host_cmd = tokio::process::Command::new(real_exe);
+    host_cmd
+        .args(args)
+        .env("AGENTMUX_BACKEND_WS", &srv.ws_endpoint)
+        .env("AGENTMUX_BACKEND_WEB", &srv.web_endpoint)
+        .env("AGENTMUX_BACKEND_PID", srv.pid.to_string())
+        .env("AGENTMUX_PENDING_MIGRATIONS", srv.pending_migrations.to_string())
+        .env("AGENTMUX_AUTH_KEY", &srv.auth_key)
+        .env("AGENTMUX_INSTANCE_ID", &srv.instance_id)
+        // Parent-identity stamp: our pid == the host's getppid (we spawn it
+        // directly). A dev-build host normally ignores AGENTMUX_BACKEND_WS
+        // (it could be a stale value inherited from a parent agentmux pane);
+        // this lets the host verify the hand-off is genuinely ours THIS run
+        // and adopt our launcher-owned srv instead of double-spawning. See
+        // agentmux-cef/src/main.rs::launcher_is_genuine_parent.
+        .env("AGENTMUX_LAUNCHER_PID", std::process::id().to_string())
+        .envs(host_env.iter().cloned())
+        // We reap children ourselves on shutdown (SIGTERM, then SIGKILL
+        // backstop) — kill_on_drop would SIGKILL the host the moment the
+        // Child is dropped, robbing CEF of the chance to reap its render
+        // subprocesses cleanly.
+        .kill_on_drop(false);
+    if disable_gpu {
+        host_cmd.arg("--disable-gpu");
+    }
+    // Linux process-tree reap: PR_SET_PDEATHSIG asks the kernel to SIGKILL
+    // this child the moment its parent (the launcher) dies, even abnormally
+    // (e.g. launcher panic, OOM, SIGKILL from outside). This is the Linux
+    // analogue of Windows' Job Object KILL_ON_JOB_CLOSE that A0 wires up
+    // (SPEC_LAUNCHER_LINUX_PACKAGED_AND_SPLASH_2026_06_05 §3 step 3 and
+    // §A1.4). Without it, a launcher crash orphans the CEF host onto PID 1
+    // and the user is left with a zombie tree until manual cleanup.
+    //
+    // The closure runs between fork() and execve() in the child; per
+    // POSIX it MUST NOT allocate, take locks, or call async-signal-unsafe
+    // functions. `prctl(2)` is async-signal-safe. macOS lacks prctl, so
+    // this is gated to Linux; macOS uses NSTask's auto-reap behavior.
+    #[cfg(target_os = "linux")]
+    {
+        use std::os::unix::process::CommandExt as _;
+        // Safety: prctl is async-signal-safe. No allocation, no syscalls
+        // beyond prctl itself. Errors from prctl are non-fatal — the host
+        // still spawns; we just lose the auto-reap guarantee.
+        unsafe {
+            host_cmd.pre_exec(|| {
+                libc::prctl(libc::PR_SET_PDEATHSIG, libc::SIGKILL, 0, 0, 0);
+                Ok(())
+            });
+        }
+    }
+    match host_cmd.spawn() {
+        Ok(c) => {
+            let pid = c.id().unwrap_or(0);
+            log(&format!("spawned CEF host pid={} (unix)", pid));
+            Some(c)
+        }
+        Err(e) => {
+            log(&format!("failed to spawn CEF host: {}", e));
+            None
+        }
+    }
+}
+
+/// Send SIGTERM to a child so it can shut down gracefully — for the CEF
+/// host that means reaping its render subprocesses (a tokio `start_kill`
+/// SIGKILL would orphan them); for srv it's a clean shutdown. No-op if the
+/// child has already exited (its pid may have been reaped). Best-effort:
+/// the SIGKILL grace-window backstop in `run_unix` catches anything that
+/// ignores SIGTERM.
+#[cfg(not(target_os = "windows"))]
+pub(crate) fn terminate_child_gracefully(child: &tokio::process::Child) {
+    if let Some(pid) = child.id() {
+        // SAFETY: kill(2) with a process-scoped pid + a constant signal —
+        // no memory is touched. A stale pid just returns ESRCH (ignored).
+        unsafe {
+            libc::kill(pid as libc::pid_t, libc::SIGTERM);
+        }
+    }
+}
+
+/// Resume the (single) main thread of a CREATE_SUSPENDED process.
+///
+/// Walks a Toolhelp32 thread snapshot to find the one thread belonging
+/// to `pid` (a freshly-spawned suspended process has only its main
+/// thread), opens it with THREAD_SUSPEND_RESUME, and ResumeThread's it.
+///
+/// Errors come from snapshot creation, OpenThread, or ResumeThread
+/// returning `(DWORD)-1`. A `ResumeThread` return of 0 means the thread
+/// was already running (impossible if the process was just created
+/// suspended) — treated as success.
+#[cfg(target_os = "windows")]
+pub(crate) fn resume_main_thread(pid: u32) -> Result<(), String> {
+    use windows_sys::Win32::Foundation::{CloseHandle, INVALID_HANDLE_VALUE};
+    use windows_sys::Win32::System::Diagnostics::ToolHelp::{
+        CreateToolhelp32Snapshot, Thread32First, Thread32Next, TH32CS_SNAPTHREAD,
+        THREADENTRY32,
+    };
+    use windows_sys::Win32::System::Threading::{
+        OpenThread, ResumeThread, THREAD_SUSPEND_RESUME,
+    };
+
+    unsafe {
+        let snap = CreateToolhelp32Snapshot(TH32CS_SNAPTHREAD, 0);
+        if snap == INVALID_HANDLE_VALUE {
+            return Err("CreateToolhelp32Snapshot failed".into());
+        }
+
+        let mut entry: THREADENTRY32 = std::mem::zeroed();
+        entry.dwSize = std::mem::size_of::<THREADENTRY32>() as u32;
+
+        let mut found = false;
+        if Thread32First(snap, &mut entry) != 0 {
+            loop {
+                if entry.th32OwnerProcessID == pid {
+                    let thread = OpenThread(THREAD_SUSPEND_RESUME, 0, entry.th32ThreadID);
+                    if !thread.is_null() {
+                        let prev = ResumeThread(thread);
+                        CloseHandle(thread);
+                        if prev == u32::MAX {
+                            CloseHandle(snap);
+                            return Err(format!(
+                                "ResumeThread failed for tid={}",
+                                entry.th32ThreadID
+                            ));
+                        }
+                        found = true;
+                        break;
+                    }
+                }
+                entry.dwSize = std::mem::size_of::<THREADENTRY32>() as u32;
+                if Thread32Next(snap, &mut entry) == 0 {
+                    break;
+                }
+            }
+        }
+
+        CloseHandle(snap);
+        if !found {
+            return Err(format!("no thread found for pid={}", pid));
+        }
+        Ok(())
+    }
+}
