+++ agentmux-launcher/src/job_object.rs
@@ -1,0 +1,51 @@
+// Copyright 2025-2026, AgentMux Corp.
+// SPDX-License-Identifier: Apache-2.0
+
+/// Owns a Windows Job Object handle. CloseHandle on drop. The job's
+/// `JOB_OBJECT_LIMIT_KILL_ON_JOB_CLOSE` flag means closing the last handle
+/// terminates every assigned process — which is what we want as a backstop
+/// if this launcher dies abruptly.
+#[cfg(target_os = "windows")]
+pub(crate) struct JobHandle(pub(crate) windows_sys::Win32::Foundation::HANDLE);
+
+#[cfg(target_os = "windows")]
+unsafe impl Send for JobHandle {}
+
+#[cfg(target_os = "windows")]
+impl Drop for JobHandle {
+    fn drop(&mut self) {
+        if !self.0.is_null() {
+            unsafe {
+                windows_sys::Win32::Foundation::CloseHandle(self.0);
+            }
+        }
+    }
+}
+
+/// Create a Job Object J0 with `KILL_ON_JOB_CLOSE`. Caller assigns
+/// processes to it via `srv_spawner::assign_pid_to_job(pid, job)`.
+#[cfg(target_os = "windows")]
+pub(crate) fn create_job_object() -> Result<windows_sys::Win32::Foundation::HANDLE, String> {
+    use windows_sys::Win32::Foundation::CloseHandle;
+    use windows_sys::Win32::System::JobObjects::*;
+
+    unsafe {
+        let job = CreateJobObjectW(std::ptr::null(), std::ptr::null());
+        if job.is_null() {
+            return Err("CreateJobObjectW returned null".into());
+        }
+        let mut info: JOBOBJECT_EXTENDED_LIMIT_INFORMATION = std::mem::zeroed();
+        info.BasicLimitInformation.LimitFlags = JOB_OBJECT_LIMIT_KILL_ON_JOB_CLOSE;
+        let ok = SetInformationJobObject(
+            job,
+            JobObjectExtendedLimitInformation,
+            &info as *const _ as *const std::ffi::c_void,
+            std::mem::size_of::<JOBOBJECT_EXTENDED_LIMIT_INFORMATION>() as u32,
+        );
+        if ok == 0 {
+            CloseHandle(job);
+            return Err("SetInformationJobObject failed".into());
+        }
+        Ok(job)
+    }
+}
