+++ agentmux-launcher/src/logging.rs
@@ -1,0 +1,33 @@
+// Copyright 2025-2026, AgentMux Corp.
+// SPDX-License-Identifier: Apache-2.0
+
+/// Append a timestamped line to ~/.agentmux/logs/agentmux-launcher.log.
+/// Best-effort — silently no-ops if the log dir doesn't exist yet.
+pub(crate) fn log(msg: &str) {
+    let log_dir = dirs_fallback_home().join(".agentmux").join("logs");
+    let _ = std::fs::create_dir_all(&log_dir);
+    let path = log_dir.join("agentmux-launcher.log");
+    if let Ok(mut f) = std::fs::OpenOptions::new()
+        .create(true)
+        .append(true)
+        .open(&path)
+    {
+        use std::io::Write;
+        let secs = std::time::SystemTime::now()
+            .duration_since(std::time::UNIX_EPOCH)
+            .map(|d| d.as_secs())
+            .unwrap_or(0);
+        let _ = writeln!(f, "[{}] v{} {}", secs, env!("CARGO_PKG_VERSION"), msg);
+    }
+}
+
+/// Home dir without depending on `dirs` for THIS specific lookup.
+/// Kept to avoid a dirs dep cycle from log() — log() is called from
+/// data_dir::resolve_paths via failure paths, and we want it to work
+/// even if `dirs` itself is mid-failure.
+pub(crate) fn dirs_fallback_home() -> std::path::PathBuf {
+    std::env::var("USERPROFILE")
+        .or_else(|_| std::env::var("HOME"))
+        .map(std::path::PathBuf::from)
+        .unwrap_or_else(|_| std::path::PathBuf::from("."))
+}
